:m: - Parented UIGradient to its intended instance.
This was unobvious, but I somehow parented a UIGradient instance to itself! This has been fixed, however.

Additionally, export status is now shown in the main "message box" while saving things to JSON files using the Rewcreation API, and RH's confirmation dialog disappears so the user isn't left wondering if the script is doing anything.