Added WIP dorm furniture support!
It isn't complete (since it needs to check the dorm store GUI to grab items' names when "PreviewToggles" isn't visible, but with these changes, the player should be able to claim a dorm, enter it, then tap the CHECK button to set up an event connection, which will lock onto any items they spawn into the room, allowing them to save them to Clothing Captures like anything else this script can save.

Make sure the item shows the PreviewToggles GUI (the GUI that lets you select a toggle, optional pieces, and variations on the left side of the screen). Tap the up/down buttons near the top-right corner of the remote to show it. If you don't, the script could error or fail to find the item's name.