:rage3: Found cause of incomplete initialization!
Yes, I was carefully looking at this script on my phone a bit ago and as I scrolled, I saw the code that reacts to the clothing customization sidebar (dis)appearing, which was used to instruct the user to equip another accessory to save it too.

Last night's dorm furniture support updates were done quickly, and I naturally (due to my AWFUL MEMORY) skipped over this reference to GUI_Base.PreviewTogglesFrame. Post-update, GUI_Base IS a reference to that Frame, since both the furniture and clothing customization GUIs have a copy of it.

I may rewrite and move this function into the validation function eventually, so the connection is always assigned to the "active" GUI/Frame instead of just the clothing one, but for now, my pink (and  hard to read) GUI won't give suggestions when you close that sidebar.