Ⓜ️ - Set FinalItem "nil" bug and fixed cam. script ref.
The camera script was still referencing the 3D preview (ViewportFrame)'s original name, "Preview", which was changed to "PreviewModel" before I added the 2D GUI preview pane and changed how the placeholder text that appears when neither of those are visible works.

Also, I tried to reference FinalItem's ClassName property right before assigning the clone to that variable, which always led to an "error while saving" on line 1716. Oops...