Renamed 3D preview cam. script.
This file's original filename was a placeholder while I was testing and fixing the main script, as I didn't know if any of my code would do what it's meant to.

Now that it has successfully saved a couple characters, I'm feeling more confident about RH Accessory Preserver v2, and I believe it will replace v1 after I add the missing ScanTypes and SaveTypes. The new version is a worthy successor to that less modular pink box, and it's up to my current standards.