Changed how EquippedStorage ref. is assigned in CharaRH SaveType.
For some reason, it seems to be trickier than it usually is to get a reference to the EquippedStorage workspace folder (where players' clothes, wings, and other stuff is placed). Instead of using the old script's function, I'm going to try getting a reference just before doing the clothing scan. Will this finally fix this...quirk? I don't know.