💾Added the RHItem SaveType!
With this SaveType's addition, the second chunk of old (RHAPv1) code has been imported into this new and improved script, since it was needed for this SaveType.

If it works in-game, the new script should be able to preview and save most clothing (other than heels, as those can replace characters' R15 leg parts).

Speaking of "previewing", I found a small typo in the camera script, which stopped it from initializing before this commit.

Also, the code used by the CharaRH ScanType has been split off into a new function, which uses a function from my game (Project Magical Mary), IterateChildHierarchy() (which compacts the code visually), then returns references to the player's EquippedStorage sub-folders in a neat, small dictionary entry. This will be reused for the HeelsRH SaveType once that's implemented.