Disabled pinch gesture and added hacky "default" cam. view.
I have to say, I'm more proud of the main script than this "secondary" one. It barely does what I intended it to, and that includes the camera, which doesn't orbit around models correctly, so most characters are clipping into or behind the camera.

I haven't disabled it, but when any new model is added to the 3D viewport, the camera view it tries to go to is immediately overridden by a basic, psuedo-isometric view, which points down at the origin from 15 studs away on all three axises.