Removed RH reference-getting function.
This may seem like  a regression at first glance, but from what I could (barely) tell, this function didn't update the dictionary it was modifying correctly...

I don't understand why it wasn't working, but it was just a failed experiment in the end. Now, RH-specific types get these references as-needed, and it doesn't feel slower than it should be, so nothing was lost.