💾Added new HeelsRH SaveType!
It's finally here (maybe)! The next RH SaveType has been added to the script, which should be able to save both modern and legacy "heels" from Royale High!

With this commit, I've also enabled "anonymization" code. When any RH model is saved, the character's skin tone is changed to a mannequin-esque light grey color, and if their whole body is saved (for modern shoes), their face and name are also edited to be more generic.

This isn't just done so any clothes I save in the future won't have Poppy's tannish skin tone. In a way, I'm also doing this to remove my username from saves. If I or someone else save clothes from RH and they get distributed, it's important that our username isn't included in the file.

When two of my accounts got banned from Royale High, it was because one of their usernames briefly appeared in a video. If that's all it takes to lose everything until 2050, I'm not taking any chances. Once furniture can be saved, the "owner ID" tag will be censored in captures for the same reason.

This shouldn't be a necessity, but Royale High's developers aren't messing around, and I won't let them destroy another one of my accounts. 😡