💾Added CharaRH SaveType (in just a day)!
Can you believe it? I can't! The next Royale High-specific SaveType is already in my handy-dandy... 📓note--erm, 📃SCRIPT! If I wrote the right code, it should be able to save full character models (with all of their equipped clothes) in a pretty low number of code lines!

With this addition, all RH clothes and characters can be saved, preserved, or...whatever you want to do with these virtual, 3D models. This in itself is a milestone, but it isn't the final one; There are still dorms (the DormRH SaveType) to save, but I don't believe it'll be too hard to add support for those soon.

This commit also corrects the placeholder "try on" BillboardGui's name, so FixRHCharacter() will know what to look for when removing all unwanted overhead icons. I've also hidden the 3D scene, since it was mainly clipping through models.