Ⓜ️ - Added extra parent checks in FixRHCharacter().
It's always a good idea to validate Instance references before trying to access their properties, children, or other related data, just in case one doesn't exist, or somehow is "parent-less". I'm not sure how a DESCENDANT of an instance couldn't have a parent (since if anything, it would be the root/base level Model I'm scanning), but based on my tests last night, that's somehow possible in this scenario.